backend/requirements.txt
django==5.1.5
django-tenants==3.7.0
psycopg2-binary==2.9.10
djangorestframework==3.15.2
django-cors-headers==4.6.0
gunicorn==23.0.0
python-dotenv==1.0.1
drf-spectacular==0.27.2
